build: Make startup script ASCII-only
It seems to be a limitation of setuptools, that your startup script
cannot be in an encoding other than that of the environment. So if you
install Hotdoc with `pip install -e` or `setup.py develop` then try to
run it with LANG=C, you'll get a UnicodeDecodeError.

Removing the copyright signs from the startup script makes it work again.