git-upload: Allow a subdirectory in --git-upload-repository
--git-upload-local-repo has the ability to accept a subdirectory, which
allows you to commit different versions of your documentation to one
repository, e.g. stable and master.

We want to allow this for --git-upload-repository in the same way. In
order to do that, we have to be able to separate the repository URL from
the path, which we do by testing with `git ls-remote`.