search extension: do not hard code limits for id lenths.
This was a hack to let the output folder be tarrable with
old versions of tar, but that's not the correct approach,
the maximum size should be passed as an argument to the
search extension.

For now, users of hotdoc that wish to tarball the output
folder need to use a modern tar implementation.