setup: search for 'gmake' in addition to 'make'
hotdoc/hotdoc_bootstrap_theme/Makefile uses non-standard syntaxes which
are specific to GNU make. On platforms which uses BSD make as the
default make command, it causes the installation to fail because
Makefile cannot be parsed. 'gmake' is a common name for GNU make on
systems which don't use GNU make as the default make command, so
searching for 'gmake' first is much likely to find the correct make.