gtkdoc_links: fix computation of symbol name
gtk-doc uses a strange pattern to store the names of signals and
properties in devhelp files, and hotdoc follows it, but not all
tools are as convoluted. As gtk-doc doesn't specify itself in the
author root attribute, we are reduced to some silly heuristics,
which should at least keep us safe.

Fixes #176