drop dependency on xdg
It's trivially implemented in a couple lines of os.getenv, and the xdg
module clashes with the *other* xdg module.

https://pypi.org/project/xdg/
https://pypi.org/project/pyxdg/

The former is a relatively recent package that ONLY provides the XDG
base directory specification, but calls itself "xdg".

The latter is Freedesktop's own module that has been around for a lot
longer, and provides the base directory variables in the submodule
xdg.BaseDirectory.*, in addition to providing code relevant to several
other xdg standards.

The existence of the former is confusing and redundant, and the use of
either one is a bit overkill; the spec is not complicated, let's just
provide it ourselves.