symbols: do not list aliases as children
Commit 9d8592fe1536ef11e0de270380916bd722902039 added these as children to have
their links resolved, but in present situation this actually leads to unresolved
links.

As they are listed as children, Symbol.resolve_symbols will also visit alias
children, and create an uninformative Link which is then added to LinkResolver.
This alias link is never resolved, i.e. provided with a title or ref, though
LinkResolver.get_named_link will return that Link in stead of performing a db
lookup using the alias id which would return the fully resolved link of the
actual non-aliased symbol.

Not marking aliases as children ensures the latter db lookup ends up with
the actual resolved Link.